refactor(codegen): carve collection lowering out of llvm.rs and add a size ratchet
Relocate the Vec / HashMap / HashSet / bytes / string collection
direct-call lowering (owned-element and layout-descriptor runtime
declaration + lowering, the option-returning collection accessors, and
the collection/bytes constructor lowering) from llvm.rs into
`crate::layout`, alongside the named-type layout machinery it already
owns. Callers in llvm/runtime_abi/thunks import the entry points from
`crate::layout`.

Add a structural line-count ratchet asserting `src/llvm.rs` stays under
a hard ceiling (at or below 45,000 lines) so the coordinating core
cannot regrow into a god-module; when it trips, the fix is to carve a
concern cluster out, proven byte-identical by `make ll-diff`.

Pure relocation: every emitted `.ll` is byte-identical (ll-diff verified
byte-identical native + wasm32). No IR change.
